Make Android source tree optional for import.
I'd added error checking to this when I removed the lunch step from
updater.sh, but tests run in CI won't have a source tree available, so
there was no way to require that at import time.

To restore the previous behavior, we can either remove the error
checking, or make defer that check until the value is used. This patch
does the latter. It'll give a clearer error message in the cases where
it matters, and will avoid causing problems when it's not needed.

So far none of the tests need this property. If that ever changes,
those tests will either need to mock this function, or perhaps a
larger refactor is needed where this path is an input to the updater
rather than inferred.

Bug: http://b/277586545
Test: atest
Test: pytest test_git_utils
Test: ./updater.sh check libxml2
Test: treehugger
Change-Id: Iee9bfbf913a1b4d28107db2e8f031fa38751a21a