Change method of passing target specific flags
This CL removes the ill-advised custom patch that attempted to add
support for passing target-specific flags and instead uses the
pre-existing, though hidden, method of doing so.

In addition, a build.rs file that invoked a nested instance of Cargo
is patched to properly pass a `--target` argument, thus ensuring the
right RUSTFLAG values are picked up.

Test: ./build.py --llvm-linkage shared --gc-sections --host-only --host x86_64-unknown-linux-musl
Bug: 278103624
Change-Id: Ic119ec7f7de488093f7a2ee23d11cd2afc6cb3a5