Temporarily enable traced for testing
The CDD doesn't enforce traced to be enabled (`persist.traced.enable`)
by default in all device types. In fact, only handheld is required to
have do to so.

Originally, the test assumes that traced is enabled by default, which
may not be true for all device types.

Improve that by temporarily enabling traced if it's not enabled, run
the tests and disable traced again if enabled by the test case.

MUST_SLEEP to allow statsd <-> traced <-> traced_probes communication to happen

Bug: 270535557
Test: atest CtsStatsdHostTestCases
Change-Id: Ie7de101975a0265de2c4a1dadacdfb3618c2c828