bazel: remove shallow_since from git deps fetching as it's likely counterproductive
Historically, shallow_since was supposed to help avoid a full repository
clone when only a single commit is necessary (apparently it's not as
simple as clone --depth=1 for commits that aren't a branch or a tag).

But the option has known edge cases and is now discouraged:
https://github.com/bazelbuild/bazel/pull/17465.

Further, as far as I can tell individual-commit fetches are properly
supported under v2 git protocol. I can't figure out how to properly
trace the git commands that bazel issues, but a manual v2 fetch against
the AOSP gerrit succeeded (per GIT_TRACE_PACKET), so I hope that bazel
will end up on v2 as well.

So my plan is to submit this and see if CI fetch flakes go away. If not
or we see evidence of full-repo clone fallback kicking in, I'll
investigate further.

Change-Id: I9d6fc5f373a84fdc434b841dbc0575fe3123dd5b