Fix test bug when log filenames collide
GcovKernelCodeCoverageCollectorTest has an internal LogFileSaver that
mocks the saving of artifacts during test run. It used a Map<> to keep
track of the files keyed by their filename. However, because of this
if two or more files had the same name each would overwrite the previous.
In actual on device testing with files saved to disk this allowed by the
framework adding a unique ID to each filename. Therefore, to get
comparable behavior we changed GcovKernelCodeCoverageCollectorTest to
use a MultiMap instead.

Test: GcovKernelCodeCoverageCollectorTest
Bug: 275565613
Change-Id: If7f59dee482aaaf2ebd7980391ccae4a2e94bcf8
Signed-off-by: Joe Fradley <joefradley@google.com>