Add support for async task tracing
The CloseableTraceScope facility does not work in submitted tasks
because the active trace context of the root invocation is lost. This
change adds an ExecutorService implementation that transparently
propagates the context to submitted tasks. This makes it possible to add
tracing logic in asynchronous code that uses the Executor framework.

Test: m tradefed tradefed-tests && ./javatests/run_tradefed_tests.sh
Test: m tradefed tradefed-tests && ./javatests/run_tradefed_tests.sh --class com.android.tradefed.invoker.tracing.TracePropagatingExecutorServiceTest
Bug: 276367158
Change-Id: I19613ffa1af8ea2ce41a3358664bd58b51ed8122