[gwp_asan] Employ EXPECT_DEATH for zxtest compatibility
Employ a similar tactic introduced by https://reviews.llvm.org/D94362
for gwp_asan tests. zxtest `ASSERT_DEATH` syntax differs from gtest in
that it expects a lambda.

zxtest does not have `EXPECT_DEATH`, so it introduced for Fuchsia builds
and wraps the expression with a lambda to create a compatible syntax
between zxtest and gtest for death tests.

An example of where this compatiblity is needed is in
`never_allocated.cpp`.

Differential Revision: https://reviews.llvm.org/D147189

GitOrigin-RevId: f2a1726d2199150918430a84f29cf8835bfe8e5a
Change-Id: I36a4ddde3fd55ae61bf731dc439762d7157dfa7d