Add better diagnostics for missing Java files.
Vogar can't really tell which of its arguments are Java source
files and which are classes somewhere along the classpath.
However IFF we get to the point of creating class jars AND our
supposed class name looks like a Java source file AND it doesn't
exist THEN we can pretty safely assume it's a missing source file.

Fixes: 239548304
Test: manual vogar runs
Change-Id: I8b98f6d263030f9ff304e4ba653fabf3689e8da7