Autoupdate of 6 submodules
*external/rust/crates/ciborium updated to de1c4e77e208abe7a1d53d9d74a50942cbefe7d8
Create no_std builds of ciborium and dependencies

pvmfw, which is no_std (but with alloc) needs to be able to read &
write CBOR (for the DICE chain), so we need a no_std version of
ciborium.

Bug: 266172411
Test: atest MicrodroidTests
Change-Id: Ia8bae9fffd7c2b98a623153c02392a81adf24fe3

*external/rust/crates/ciborium-io updated to 6cb80b001e92d1ce3be758caa43a9db787070562
Create no_std builds of ciborium and dependencies

pvmfw, which is no_std (but with alloc) needs to be able to read &
write CBOR (for the DICE chain), so we need a no_std version of
ciborium.

Bug: 266172411
Test: atest MicrodroidTests
Change-Id: I3d2a1f4185a6ec41ba8205f2306e37e6823998b6

*external/rust/crates/ciborium-ll updated to 7437881bf5d5bbc017315cc62c04640c3d06d409
Create no_std builds of ciborium and dependencies

pvmfw, which is no_std (but with alloc) needs to be able to read &
write CBOR (for the DICE chain), so we need a no_std version of
ciborium.

Bug: 266172411
Test: atest MicrodroidTests
Change-Id: Ifeac3712d1516264bd39ba2a9caede4e9f6aeaec

*external/rust/crates/half updated to 03842cff6fcb5d00e38a6744c6e49aa7f688f965
Create no_std builds of ciborium and dependencies

pvmfw, which is no_std (but with alloc) needs to be able to read &
write CBOR (for the DICE chain), so we need a no_std version of
ciborium.

Bug: 266172411
Test: atest MicrodroidTests
Change-Id: I8117e7ce2bbcdc2def994196693efa51399fddac

*external/rust/crates/serde updated to f7d51f4fa182c7897cb2a53c49563271fdefb6ad
Create no_std builds of ciborium and dependencies

pvmfw, which is no_std (but with alloc) needs to be able to read &
write CBOR (for the DICE chain), so we need a no_std version of
ciborium.

Bug: 266172411
Test: atest MicrodroidTests
Change-Id: Iee89507487203e29ad7076276e13f29d4d3caabf

*packages/modules/Virtualization updated to abc93ebfed31dd42722c397b37800cfa05519d4f
Merge "Add Ciborium to pvmfw"