Autoupdate of 3 submodules
*device/generic/arm64 updated to 04bf0983aa0ff1aa5892130f08472c1227769182
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I7fddc45b33e0bdbaaaadc3f7cbafcb4965343886

*device/generic/x86_64 updated to 5daaac4d2834c63238e68af21e3c32a1c5466a6a
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: Id061a7a22b185b6d0afb2bc73ff0717132cf95a0

*device/linaro/hikey updated to b82a3bb29b6e3cfc802fa10426e5f88ab9515806
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I89e0cd7c159101f2fd96c88f4dcdf58473cf9ed2