Autoupdate of 10 submodules
*device/amlogic/yukawa updated to d06617d6d331446e5d609eca48158a0b66b43ad2
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I73df5b80ddf8fa9b3a9bb150e68bfb0a06aa9593

*device/generic/arm64 updated to 04bf0983aa0ff1aa5892130f08472c1227769182
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I7fddc45b33e0bdbaaaadc3f7cbafcb4965343886

*device/generic/trusty updated to 5e9ea6f3bf8d158d3424a8042339aa531cf9c0a1
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I62125f0ce7e44c1a9b53f52da5c0f360bead85ca

*device/generic/uml updated to 672257b0995d672a4517bcdd6b782d2056e7146b
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I03141d6413776b125abd94d6bcdad361094a19e0

*device/generic/x86_64 updated to 5daaac4d2834c63238e68af21e3c32a1c5466a6a
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: Id061a7a22b185b6d0afb2bc73ff0717132cf95a0

*device/google/cuttlefish updated to 4548b2df4e37896dd0d31859b3204f32b7c5238b
Merge "Remove 64-bit binder ABI setting."
*device/linaro/dragonboard updated to 86f25c1d4b3ca9969524abd127bb3e86a5f71e42
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: Ib3b6f866c966f34aa19cedf47b716a0558131df8

*device/linaro/hikey updated to b82a3bb29b6e3cfc802fa10426e5f88ab9515806
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I89e0cd7c159101f2fd96c88f4dcdf58473cf9ed2

*device/linaro/poplar updated to 0a95a7b3e2c18bde907553c8e9b3eb38295219c8
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I9302b429eb2a8bd8589c91add108f814fa7cc387

*device/ti/beagle_x15 updated to c25c12f0ffb5b34c439c46fe4a1873104d37c855
Remove 64-bit binder ABI setting.

We are only using the 64-bit ABI now, and this
setting is being removed.

Bug: 232423610
Test: TH
Change-Id: I93e8ed81f8746ded4b02d362b7a1c5533738f381