Autoupdate of 3 submodules
*frameworks/native updated to d5ab777ad62e4cf43cfeddd1b1d5f68c625074e7
Merge "libbinder: replace flaky tests with real coverage"
*system/sepolicy updated to d8b05e70bf3ca3002d1bdda3376228ebe74005d5
Merge "aidl_lazy_test: additional service context"
*system/tools/aidl updated to 2aa12d8874054c53cb0eb754124799dc08fb1988
aidl_lazy_test: test service with clients quit

A lazy service shouldn't quit when it has clients, but
sometimes it needs to, such as when the device is
shutting down, so we test that it works.

In Android U, I broke this behavior, and it was caught
by other tests. However, now we have test support
for this directly in aidl_lazy_test.

Bug: 279301793
Bug: 278337172
Bug: 277886514
Bug: 276536663
Bug: 278117892
Test: aidl_lazy_test
Change-Id: I5e6a8e08973b4494979f4d9e2fc52fe6cca99ddb