Fix flaky tests.
There is a small chance that doing the attach at the wrong time
causes the process to die, or you can't attach at all. On device,
sometimes a process gets stuck when doing the attach and never
comes back.

In order to avoid these cases, modify the code to always retry running
the process if these failures are detected.

There is still one failure in UnwindTest.thread_unwind that is a
race condition in the test. I'll fix that in a follow-up.

Bug: 273963946

Test: Ran the RegsRemoteTests suite hundreds of thousands of times
Test: on host without a failure.
Test: Ran the full suite thousands of times on host with only one
Test: failure.
Test: Ran the full suite 924 times on device with only one failure.
Test: Ran the full suite on a MTE enabled device for more than 3500
Test: times without failure.

Change-Id: I346f27e7138a0ca5eab4a1fcb9e0dfe32f555170