Simplify bool parsing
I find this version easier to read. In addition, using a pattern gives
the compiler a tiny bit of freedom in how to implement this.

For fun, I looked at the compiler output. In this case, the pattern
matching is implemented it via a series of direct ‘cmp’ calls[1]
whereas the ‘contains’ version had a call to ‘slice::contains’.

[1]: https://rust.godbolt.org/z/5baxKEjE8

Test: atest librustutils_test
Change-Id: I11c6e857a3eb84794798cebce96d1be3d671bb8f