Fix race in activating sharedlibs APEXes
ActivatePackageImpl() should be thread-safe, but the current logic to
SetLatest() shows TOCTTOU problem.

To fix it, we no longer keep "latest" flag. MountedApexDatabase keeps
MountedApexData along with version info. Operation for the lastest mount
is now guarded by the DB's mutex. (Bind-mounting the latest mount to the
active mount point).

Bug: 279652747
Test: ApexTestCases
Change-Id: Ic0b1d4f888e316aeb5c30b57c941913741761505