bpfloader: unconditionally set bpf_jit_kallsyms=1
This should now be safe on Android V+
(we should have aged out of kernels that don't support this),
as we're not really considering R->S->T->U->V 4 OS version upgrade path
supported (without at least some work on the part of vendor/oem).

Note that 4.14 was the minimal version supported for U,
so presumably the minimal version for V is 4.19,
which happens to also be the only version < 5.4 (ie. where
this isn't already enforced).

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I22487bf54d8f8e235b4dde3cdde5c3c2c1ccf9e8