zram: move to generic-tests
Although strongly encouraged, a zram-based swap space may not be used
in all devices. One example is automotive, where the usage of
suspend-to-disk (aka hibernate) for fast boots is common but
incompatible with such a swap space.

Since there is no CDD and/or VSR requirements around the zram-based
swap space and this test simply checks for the usage of such a swap
space and not the functionality, move the zram-related tests out of
vts to general-tests.

Bug: 273839666
Change-Id: I2e5cf0e866af10133933a98e327b023ab944d595