Autoupdate of 2 submodules
*art updated to 32645f219cb3d4588c1aefcf9c7db944aef6cd45
Remove the codegen property

The codegen property was very similar to the targets property, except:
  - on host it combined all the properties from all the arches
  - it would apply the properties from each arch branch if the device
    supported it, with the exception of always including a 32bit branch
    if the 64bit branch was included.

For the most part, this could be expressed by using the regular `target`
property in soong. However, the `target` branches are selected based
on if the module itself is being compiled for 32bit/64bit/a particular
arch, not if the device supports it or not. This means that after this
change, compiling a module for 32 bit on a device that supports both
64 and 32 bit will not have the 64 bit flags. Note that the reverse does
not work the same way, the 32 bit flags are manually copied onto the 64
bit `target` branch to maintain the plugin's old behavior of always
including 32bit code on 64bit.

Bug: 247785938
Test: Presubmits (because of the behavior change described above, the ninja files do have differences)

Change-Id: I23f585c2a2ad2cbe9b3e212097aef88d63c5394b

*external/vixl updated to e9748ade075bda078a7f3e7041d0afdd5f07b08e
Remove usages of art_cc_defaults

Replace the codegen property with the target property.

Bug: 247785938
Test: See other cl in the topic
Change-Id: I1dccba261ae5ab0e19cc4ae4647b0ec02e6c2427