Use CLASS retention if explicitly specified
This avoids retrieving the default retention which will fail when
called on a TextClassItem as the default retention differs between
kotlin and Java but the Text codebase does not differentiate between
kotlin and Java. However, that does not matter in this case as the text
files always specify a retention policy.

Bug: 281035030
Test: ./gradlew
      m droid
Change-Id: I706a0f48167e96746f86635fdf6a39687f52be0d