Autoupdate of 2 submodules
*bionic updated to 983292a9724604046fbd4e4002c0bc8d8bc09347
Merge "Update to v6.3 kernel headers."
*hardware/google/graphics/common updated to a13b136e4a7d43949bf9118c953df92eb64db394
Update to handle kernel 6.3.0 headers.

It looks like a couple of defines are incorrect. Change them to
be the same as in the kernel. These defines are not actually
used in the code yet, so this does not cause any changes in real
code.

Test: Builds.
Change-Id: I33785b0ff0245a4ba47c16517a90ca31c8d9e984