Autoupdate of 2 submodules
*external/selinux updated to e1c842285bb4d94b40488471ea6c72688f3ff5f7
Add applySdkSandboxNextRestrictions flag

seapp_context_lookup_internal applies a flag that is referenced in
seapp_contexts based on the seInfo string passed to it.

This enables testers to test out the set of restriction planned the
next SDK version and give feedback before we decide on the actual
restrictions for the next release.

Bug: b/270148964
Test: manual test app and adb shell ps -Z
Change-Id: I175229d135d99516dd6f38b8963d0ccc93a61a4f

*system/sepolicy updated to 1c18e1a92212662ee30b552b261fa527d7722e2e
Merge "Add canary restrictions for sdk_sandbox"