[aidl_rust_glue] Allow passing the arguments in a file
aidl_rust_glue expects to be given all the .rs files generated by
aidl_gen.  Some targets might have enough file to overflow the max size
of the kernel command line

This CL enables putting the arguments in a file, and passing the
filename, prefixed with '@', to aidl_rust_glue.

Change-Id: Icdd36b5992296d87235bdb9ee65f162725c46c2e