Autoupdate of 2 submodules
*hardware/interfaces updated to 1d23120bad86250581c33ca4f3f6c2577ff49865
Merge "Add security version to config descriptor"
*tools/security updated to 4028ff70ebf2efff570457eff229a4acd668d4f8
Add security verion to config descriptor

Parse and validate the security version in the configuration descriptor.
The field is not required so its addition is backwards compatible.
Previous VSR versions are permitted to include this field now that has
been allocated and defined.

Test: atest --host hwtrust_cli_tests hwtrust_tests libhwtrust_tests
Bug: 282205139
Change-Id: I10bf4e0727c9e1b6bb8e0b0fa6da6a69d4ea72e7