Block on joining IO threads in RunUtil
If the IO parsing threads in RunUtil don't join in the given timeout,
then all it does is print a debug message and keep going. If there are
result reporters relying on those IO threads to complete they can be put
in an undefined state if the test finishes but results are still coming
in.

Given that the behavior of RunUtil should be to run until either the
command finishes or the given timeout runs out then it should wait for
everything to complete before finishing, not just the system command
itself.

Test: presubmit
Bug: 281902802
Change-Id: I1a18492b02f1849914ae32e8323d93822dd89fdf