Always use debuggable from compiler options when JITing
We use two slightly different ways to check if we need to generate
debuggable support for JITed code or not. We usually check for
debuggable flag from the compiler options but also use
Runtime::IsJavaDebuggable in some places. While these two usually match,
it is possible to set compiler options independent of the runtime
debuggability. So always use debuggable from compiler options when
making decisions on JITed code.

Bug: 282239094
Test: art/test.py

Change-Id: Ia47f146834e0ae1aec2c5694d3d7a468be51435f