Remove extra error logs when parsing non-EROFS images
When running delta_generator, it will first attempt to parse input as
EROFS images, and try other image format it that failed. delta_generator
will print an ERROR message when parsing EROFS failed, but that should
not be regarded as an "ERROR" by developers looking at the logs. To
avoid confusion, check for EROFS magic and return silently before
performing actual parsing.

Test: th
Change-Id: I23f5fe49ca109231b85b21e1d6ffa6b97f96b9fb