Add an experimental flag to reduce build targets
Atest adds dependencies that are unnecessary for running a test which
significantly increases the build time. This change sets up the basic
feature to only build a minimal set of targets for the correct variant.

Reducing the build set reduces invocation times. For example, running
CtsViewTestCases on a clean checkout drops the build time from 701s to
433s (-38%).

Test: m atest && atest-dev atest_unittests
Test: m atest && atest-dev --minimal-build CtsViewTestCases
Test: m atest && atest-dev --minimal-build HelloWorldTests
Bug: 281058510
Change-Id: I4b055227254a95521efd5c630b3b168a12c117d1