Require the config hash
Check that the configutation hash is present and corresponds to one of
the three supported hash functions. The Open Profile for DICE requires
that the config input is explicity listed in the certificate and, since
Android uses a configuration descriptor that is hashed, this means that
the configuration hash must field be present.

Test: atest --host hwtrust_cli_tests hwtrust_tests libhwtrust_tests
Change-Id: Ia88e0f9460d284d3873a578f7b96d07d882d8b32