Wait for test network disconnection before exiting tearDown
The test received an unexpected callback because the NetworkAgent
which is registered by the previous test is not disconnected yet.
So the new test has chance to receive the network which is created
by the previous test and which is going to be disconnected.

Hence, this change make sure all networks has been disconnected
before exiting tearDown.

Test: atest CtsNetTestCases:NetworkScoreTest \
      --rerun-until-failure 1000
Fix: 242675267
Change-Id: Ic2451b10bec0e656d4651b5f8ee99ef437e92695
(cherry picked from commit b5878363f16e325c23435a381f4c366eed6f5c49)