Convert xsdcCppRule to a sandboxed action
Sandboxing helps ensure that all dependencies need to be declared
explicitly in Android.bp files. This helps migration to Bazel which
enforces per-action sandboxing (atleast out of the box).

The sandboxed Soong action does not need to emit a depfile since the implicit
include deps are now declared explicitly in Android.bp files

Test: m # passes
Test: Built ApexServiceTestCases which uses an xsd_config # no diff
Change-Id: Ife9d1119be34442d8e5ac0d20343477549ffa8f5