Move FinishedSnapshotWrites call to filesystem verifier
FinishedSnapshotWrites() must be called before slot switch and after
finish writing data to the target partition. It also requires all target
partitions to be mapped when calling. Priori to this CL,
PostinstallRunnerAction is responsible for calling
FinishedSnapshotWrites(). However, postinstall action has complicated
logic to decide if partitions should be mapped/unmapped. So it's
difficult to ensure that partitions are mapped before calling
FinishedSnapshotWrites(). To make things simpler, move the
FinishedSnapshotWrites() call to FilesystemVerifierAction.

Since FilesystemVerifierAction is responsible for performing the last
write to partitions(verity writes), this make placement make sense. We
are also guaranteed that partitions will be mapped, as
FilesystemVerifierAction need to map these partitions in order to do
verification.

Test: install VAB test w/o slot switch, call setShouldSwitchSlotOnReboot
Bug: 269397263
Change-Id: I900b529f8b4a0b604bab3a334f61ff9a51139916