Autoupdate of 2 submodules
*prebuilts/gradle-plugin updated to aced32641c628d193dd45e50ed0518c2d02bcb6b
Merge "Upgrade metalava lint prebuilts to 31.1.0-alpha07"
*tools/metalava updated to 4941f746ce1c9404ef49316863d6d963886f8da5
Fix "overrides nothing" error in PsiBasedCodebase.kt

The previous merge commit changed the definition of a couple of methods
in the `JavaRecursiveElementVisitor` that were overridden in
https://r.android.com/2521896:
* `visitCodeBlock(PsiCodeBlock?)` -> `visitCodeBlock(PsiCodeBlock)`
* `visitDocComment(PsiDocComment?)` -> `visitDocComment(PsiDocComment)`

That change caused a build failure as the overrides had a different
signature which meant they did not actually override anything. This
change corrects the signatures.

Bug: 281035030
Test: ./gradlew
      m checkapi
Change-Id: Id6d1ceebd86ab5a38e391534002dc530a3ae2fbb