dexdump: Fix confusing --help listing
The --help screen was confusing. Options "-h" does not display the file
header details but the section details. To get the dex file header to
display, you need to use "-f" which was labelled "summary information
from file header"

This CL clarifies the --help text so "-h" and "-f" cannot be confused
with bette explanation.

Test: NA
Bug: NA
Change-Id: I522d31c91938643984032f5484b35c47a71304b7