Replace custom URL builder code with Uri.Builder
Use a more standardized way of building URLs than the previous method
within RKPD. This reduces the chances of making mistakes going forwards.
Due to the presence of a `:` in the endpoints for the services, there is
still a bit of hacky clean up of the URL to do in performing a manual
string replace for the URL-encoding of `:` ("%3A").

This can be fixed once the next version of the service API is launched.

Fixes: 279500127
Test: atest <all of the Rkpd MTS test suites>
Change-Id: Ic06621621f3a77ac68fa96f58964e90d0140690e