Autoupdate of 2 submodules
*prebuilts/rust updated to 9225bc371974f970097f63e32cd67b42728a3473
Remove the rustfmt prebuilt definition

This version of the rustfmt prebuilt definition is incorrect.  The
correct version won't be able to land until the next version of the
Clang toolchain is released where the DT_SONAME values in the shared
libraries do not use versioned .so names.

See aosp/2585829 and aosp/2585130

Bug: 274972941
Test: m rust
Change-Id: Ie7530525f6fb801ce46922b29c2194cf13c7ff11

*toolchain/android_rust updated to 02d0d9dabd8b4ed692023c7343a6e26f432842d3
Merge changes from topic "rust-remove-prebuilts-1.67.1"

* changes:
  Adjust handling of stable links when updating prebuilts
  Remove artifacts from dropped prebuilts