logd_integration_test: inject state to device logs
By injecting the state to the device logs, we can see which logs
are counted as part of the test, and which logs are caused by
other tests.

The current situation is muddled because even though we use
RebootTargetPreparer here:
- cf logs are stored without drops on the host for many tests
  run in a row, so it's hard to see which parts of these logs
  correspond to this test
- RebootTargetPreparer is set to only boot before running the
  test for this test, not after. This is expected, because this
  test does not mess up device state, so we shouldn't pay the
  cost for the additional reboot.

Using these logs, the flake encountered in the bug this fixes
would have been debuggable from logs.

Fixes: 284033296
Test: logd_integration_test
Change-Id: I9cc9cf06e1efb259985c22ac920b069e7a3f1845