Add --api-overloaded-method-order
The preceding change merges in a change that alters the order of
overloaded methods in API signature files.

The current order is determined by their order within the source code,
and if the overloaded methods are in different classes then falls back
to using signatures. That makes the API signature file sensitive to
source refactorings that do not change the API which will result in
unnecessary churn in API files. The new order is better as it is
determined solely by the methods' signatures and so is not affected by
changes to the order in the source files.

Unfortunately, switching to that new order causes practically every
single API signature file in the Android build to require updating
which would be a logistical nightmare to try and update them in one go,
particularly while trying to merge in change from `metalava-main`.

So, this change adds an option to allow the sort order for overloaded
methods to be specified on the command line. By default if the option
is not specified then it defaults to the existing `source` order. The
tests default to using the new `signature` order as very few tests are
affected by it and all but one of those tests has been fixed in the
merged in change. The exception is the following test that was added in
`master` since the fork.
'''
ApiFileTest.'Test can merge API signature files with generic type classes'
'''

That test is set up to rely on the default setting of the option.

Bug: 281035030
Test: ./gradlew
      m checkapi
Change-Id: I0883c344c015b2b5c7e367feb7bf81ca087e34ff