Stable AIDL: ignore @JavaPassthrough
This also simplifies the way that @Backing annotations
are checked in order to avoid losing test coverage.
Previously, @JavaPassthrough was used in order to test
checkapi's ability to compare value arguments in
annotations.

Now, we use @Backing for that test, and the duplicated
logic implementing this same check for @Backing is removed.

Bug: 285018823
Test: aidl_unittests
Change-Id: I81981eb06f76709e9462a0cc9585f246aa4e7fc7