Temporarily enable traced for testing
The CDD doesn't enforce traced to be enabled (`persist.traced.enable`)
by default in all device types. In fact, only handheld is required to
have do to so.

Originally, the test assumes that traced is enabled by default, which
may not be true for all device types.

Improve that by temporarily enabling traced if it's not enabled, run
the tests and disable traces again if enabled by the test case.

Bug: 285215697
Change-Id: I3e93e4b93a0349fe3d9f790525ae61a0f484bc97