aidl: include generation command
AIDL is used in a lot of places now, so I went ahead
and added the commandline arguments to the top of every generated
file. This will help in Google3 and Android Studio and in
another upcoming place.

This does not add the command to the top of API dumps, and this
is intentional because these files should really only ever be
created by a build system which enforces them.

Bug: N/A
Test: mma
Test: check the only files in tets/golden_output without 'MODIFY' are
    .d files or timestamp files
Test: aidl --lang=java ./build/tests_transitive_parcel/a/IA.aidl --out . -I ./build --omit_invocation
Change-Id: If5481974aa016eaf261be8e4a63eba5363c0b63a