Create a denylist for 1:N mapping of xsd_config
The output of xsd_config in Soong are source files (.cpp/.srcjar). Not all
.xsd files generate compile-able source files though. This is not an
issue in Soong because these uncompile-able sources do not have any
rdeps.

However, this will be an issue for Bazel CI since the output of the equivalent
*_xsd_config_library are compiled binaries. To prevent this, this CL
introduces an allowlist in bp2build to skip generating these targets.

Bug: 211678537
Test: bp2build.sh
Change-Id: I44d007a2d9895442a0f37e2d77f8b327e00dee14