Fix bound space bitmap assertion
Following the refactoring of space bitmap changing how binding of
bitmaps is implemented, the assertion was no longer valid:
  I4b1040d29f9ba99ea23c51f70a06af2ffbe496d4
It was checking that the two bitmap objects were located in the same
memory area which was never correct.
Change the assertion to instead check that the bitmaps' internal storage
address is the same.

Test: Initial stages of DaCapo/Xalan run with libartd.so runtime
Change-Id: I09761cdce65e74001363ef608d36b50202d62c9d