Create GET request route for initial connection to websocket
The sample URI for GET request is the following:

/v1/websocket?name=device&kind=BLUETOOTH

The handler will check if query string is valid and if it contains
device name and chip kind before running websocket transport.

Next step is to implement upgrading to websocket server and connecting
it to the packet dispatcher.

Test: netsimd -d
Video: go/netsim-video-cl2618167
Bug: 286134569
Change-Id: Ie3f621493c7bb216809bf3d3055aa9b9541e34e3