NDK backend: allow in-process comms w/ Rust
In-process comms existed below the libbinder level, but
it wasn't implemented above that level because when the
NDK backend was implemented, there was only the NDK
backend on top of libbinder_ndk. However, since Rust
is now on top of the same API, we need specific support
here for C++<->Rust in-process interop.

b/286436233 is filed to remove as much of the auto-generated
code from generate_ndk.cpp, because it has grown too large.
Moving it C++ headers will also make it easier to compress
the code or remove templates from it. I haven't done that
here as we will want to move all of fromBinder (not just
the part that is added here).

Fixes: 286246993
Test: binderRustNdkInteropTest
Change-Id: Ib4b7a7b3d0f2ab1393aa7f9e7a8b00182061b380