results: refactor return values to include the full set of results
Rather than normalize everything into a bool early on, and then make
it hard to keep track of results across projects, and force creation
of a parallel structure to hold fixup commands, create a new results
object to hold everything and pass that back up the chain.

For now, this is just restructuring, and upload continues to behave
the same.  But this will make it much easier to change the UX.

Bug: None
Test: unittests
Change-Id: Idfa545f02f3d26a893d6cbbdef6d9a2409e72813