pre-upload: show all possible fixup commands while running
This allows people to quickly run fixup commands themselves if they
want.  This is helpful as we only offer to run the command if there
is one fixup in the latest commit.

Show these as a summary at the end so hopefully it's a little easier
for users to access if they want to copy & paste the command.

Bug: 274529930
Test: unittests
Change-Id: I75b9c196a5369874fac94e822ec87e72f804cb9a