pre-upload: fix & improve display of files
In a refactor long long ago, this line was switched from a print (which
adds a newline to the end) to a string concat, but the newline was lost
which has caused the error display to be smooshed together.

Also pass the contents through rh.shell.cmd_to_str to get more natural
output & quoting so people can copy & paste into a shell rather than
the Python stringify output for lists/tuples.

Bug: None
Test: uploaded a CL with formatting errors
Change-Id: I02acbb39acf5cf09274897086c107170c4fc8554