Fix for NDK HWASan on RISC-V
We can no longer assume that because ART is not compiled with HWASan,
HWASan is not active. There could be instrumented JNI frames, so we need
to do a runtime check for whether the HWASan runtime is loaded, via a
weak symbol.  This is the same we do for arm64.

Bug: 276930343
Change-Id: I7a7f85fc48a2e9f510dd5421ca3a55d6f0cce2c9