pre-upload: hoist exclusion logic earlier
Filter the set of hooks as soon as we have all the info we need.
If all the hooks are filtered, we can exit sooner.  Otherwise,
present a more accurate summary to users as to how many hooks
we'll actually be running.

Since we no longer need the scope in the main loop, switch to
processing the NamedTuple directly so we don't have to unpack
every field in it.

Bug: 160223496
Test: unittests
Change-Id: I63d65a9034a6d4ca6d70b158ceab7c6f5861f326