Autoupdate of 2 submodules
*packages/modules/adb updated to 65639ca242df9d40116e8e386b3ba41549e0eb4d
Faster adb unroot/root

Previously, adb unroot and root could experience up to 5 second delay
because init didn't restart a service unless it was started +5 seconds
earlier. This was a mechanism to rate limit crashing services, but adbd
(whose exit was deliberate, not a crash) was affected as well,
especially when adb root is attempted right after the boot.

Init now allows restart_period to be less than 5 and respects this if
the service is not crashing. To make use of the change, adbd now

(1) exits with status code 0 and
(2) explicitly sets restart_period to 0

Change-Id: I6c5ff908ab6a6f0db08145882cd79da1008e0d13
Test: In progress
Bug: 286061817

*system/core updated to acfc93f92427fbc132de8c76e0787d3c1387cee0
Merge "init: non-crashing service can restart immediately"