RESTRICT AUTOMERGE: Move updated fragment_density_map tests to excluded list for android T
Some vendors passed CTS 13_R3 but fail in CTS 13_R4/CTS 14_R1
This happens because the test contents have changed here:
https://github.com/KhronosGroup/VK-GL-CTS/commit/1de0a0106b4eab9c8ebb518a9e447dc66139db85

Now some devices are struggling to pass the new CTS. We need to allow this due to VSR constraints.

This CL moves the tests that cause issues from 2021 to excluded.

This CL was closed in android U, but since there's no need for 2023
list in AOSP, we just exclude these

Merged-In: Ic62f6929984636ddeaea114f41e17dcfa8fdcebe
Bug: b/285981913, b/286187826, b/286780811
Change-Id: Ib042afa0b76e11759a1ca6b04273dcff777cac88