Autoupdate of 2 submodules
*device/google/cuttlefish updated to 7d6ba31108aae165cb45759cce9f27ab6e4e100b
Merge "Disable interaction with the core scheduling feature."
*external/crosvm updated to feeb02d5252480c114edcc42537c45818e7134b7
ANDROID: Add `--core-scheduling` flag gating interaction with PR_SCHED_CORE

On Android CI we observe performance regressions when crosvm takes
advantage of the PR_SCHED_CORE feature. We are somewhat oversubscribed
on CPUs, in some cases running a Cuttlefish Android VM with --cpus=4 and
an OpenWRT VM with --cpus=1 next to some other host processes on cloud
instances with 2 cores and 4 hyperthreads. In this case we would prefer
not to lose cpu time to the scheduler blocking off hyperthreads when
either VM claims complete cores to itself.

In this case we are intending to fall back to the default state of "all
processes trust each other", mentioned under "Trust model" on
https://www.kernel.org/doc/html/next/admin-guide/hw-vuln/core-scheduling.html

`--core-scheduling` defaults to true, so there is no behavioral
change for existing users.

Bug: b/280660768
Test: `crosvm start --core-scheduling=false` from Cuttlefish launcher script
Change-Id: Id154790c16b7d9f81aff1f189468959fb5fa7259
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4602908
Reviewed-by: Frederick Mayle <fmayle@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Cody Schuffelen <schuffelen@google.com>