Explicitly set CMAKE_INSTALL_LIBDIR
GNUInstallDirs.cmake autodetects a library directory using the host
Linux distribution. On non-Debian, non-Arch distributions (e.g. RHEL),
this is "lib64". However, the toolchain build system expects all
libraries to be installed to a directory named "lib". Since some
libraries appear to be missing, the stage2 build fails with a
'Lookup of version before library is built' exception.

Override CMAKE_INSTALL_LIBDIR so the libraries are placed in the
expected location. This allows the toolchain to be built on RHEL.

Change-Id: I488d57b8fc6b0c33063486b3e25bdc5247f74262