remove apex_available properties from ot_rcp_cflags_defaults
The RCP simulation will be called by the HAL, so that the ot-rcp
is located in the vendor partition. However,
the ot_rcp_cflags_defaults contains the apex_available properties,
which causes the compile error on some Thread stack enabled
branches. This commit removes apex_available properties from the
ot_rcp_cflags_defaults to avoid the compile errors.

Bug: 286986368
Test: Build
Change-Id: Ib481c5727d72dadb0deabaacb5a547f0676d3a51