Reduce multidex checksum to single scalar value.
Change GetMultiDexChecksums to return a single value rather than
a vector. The single checksum represents the whole multidex set,
and will change if any dex file within the multidex set changes.
Fundamentally, we need a value to represent a .zip/.jar archive,
so that we can check whether our build artifacts are up to date.
We previously used a vector of CRC32 values (one per dex file),
however, one combined checksum per archive is also sufficient.

This is necessary, since the future multidex format will have
single zip entry, so we need to adjust the code to expect just
one checksum per zip file. This separates the change to own CL.

The two sides of the checksum comparison are trivially reduced:
 * Zip files are reduced by XORing the individual CRC values.
 * Likewise, checksums of already open DexFiles are XORed.

As a consequence, ClassLoader path needs to be reduced to
print only single checksum per jar file as well.

Bug: 266950186
Test: ./art/test.py -b --host
Change-Id: I848aee1e4836e87945a5172c7594e266739451e9