Use BMP mode with max window size of 4096
This change increases the window size to the maximum allowed by the
kernel, allowing out-of-order packets to be accepted beyond the current
limit of 32. At 1Gbps, this will allow for packets to be delayed by up
to 40ms without dropping.

In-order delivery of packets is generally still expected, but this
allows for packets that traverse different paths (e.g. due to different
IPv6 flow labels that are copied from inner IPv6 packets) to not get
spuriously dropped based on other network traffic

Bug: 287164181
Test: atest XfrmControllerTest XfrmControllerParameterizedTest
Test: atest NetdBinderTest
Change-Id: If784a440d416eeb126dcdf4d1a8f4e029ecc3142