Autoupdate of 2 submodules
*build/kati updated to 35e4da8ff4d86206e06296d8328b8feba971d35e
Add arbitrary metadata "tag" tracking.

Allows makefiles to set the `.KATI_TAGS` target specific variable to
annotate arbitrary metadata for the build step for use in profiling.

Bug: http://b/259130368
Test: end-to-end test tracing the time spent running cp for dist
Change-Id: I930a828cc09add1ce13f7e23142e208c9a7ca5fc

*external/ninja updated to 45d6a94d6de54494150299ff0e7137196f3f0b47
Add tags field to build.

The tags field is a string with an implementation defined format. This
allows build systems to include arbitrary metadata in the proto for
better metrics.

Bug: http://b/259130368
Test: used to trace dist cp time
Change-Id: I9f6fcae0d8e4f7152ea82d3038807f622ca92801