Use Clikt help formatting for top-level metalava command
Previously, the `MetalavaCommand` simply delegated to
`Options.getUsage()`. This change integrates the Clikt help formatting
to do the majority of the help and appends the existing `Options`
usage to the end.

It attempts to preserve the style of the previous help output but there
are some minor differences with the formatting of the help text and use
of italics for highlighting the default value.

Some manual testing from the terminal was performed to ensure correct
and consistent styling for all the ways to trigger help

Upgraded to Clikt 3.5.3 to support treating long options that start
with a single `-`, e.g. `-new_api` as an unknown argument that can be
passed through to `Options`. Previously, in 3.5.2 adding a short option
like `-h` would cause Clikt to treat options like `-new_api` as a group
of unknown short options and fail.

See https://github.com/ajalt/clikt/releases/tag/3.5.3 for more info.

Bug: 286023667
Test: ./gradlew --rerun-tasks
      ./gradlew installDist
      export PATH=/usr/lib/jvm/java-17-openjdk-amd64/bin:$PATH
      ../../out/metalava/install/metalava/bin/metalava -?
      ../../out/metalava/install/metalava/bin/metalava -h
      ../../out/metalava/install/metalava/bin/metalava
      ../../out/metalava/install/metalava/bin/metalava --fred
      ../../out/metalava/install/metalava/bin/metalava version -h
Change-Id: I75679bee21060770670d7d83dd8936a2857135c5