Add simple but extensible DSL for command tests
The existing `DriverTest` class provides support for testing almost
every option that is used by metalava but it so in a monolithic file
which is difficult to split apart. This change adds a simple DSL for
building command tests which can be extended as and when needed.

The DSL is expected to evolve over time by adding more functionality
needed by the different sub-commands being tested. However, that
functionality will be general and not coupled to a specific command.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: Id86671a38377ae686db838abe2b515b236cabb29