Extract Manifest from DefaultCodebase
The `Manifest` really has nothing to do with the `Codebase`. The
`Codebase` was simply a convenient place to use to tunnel the manifest
through to where it was needed. This change extracts it into a separate
class and migrates the `manifest` property to Clikt to allow it to
reference a `Manifest` directly.

A follow up change will remove the `Codebase.manifest` property
altogether.

Bug: 287241408
Test: ./gradlew
Change-Id: I9ce8f1cb63a842d2ee0e419aff883cb60d496fb1