vmbase: Set no_stdlibs for libvmbase
When building vmbase clients (e.g. `m pvmfw_bin`) with code changes in
vmbase, the build system masks (for some obscure reason) any vmbase
error with the following

  FAILED: [...]/libvmbase/android_arm64_armv8-2a_cortex-a55_rlib_rlib-std/libvmbase.rlib.clippy
  error[E0152]: found duplicate lang item `panic_impl`
    --> packages/modules/Virtualization/vmbase/src/lib.rs:43:1
     |
  43 | fn panic(info: &PanicInfo) -> ! {
     | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
     |
     = note: the lang item is first defined in crate `std`.
     = note: first definition in `std` loaded from [...]/libstd.rlib
     = note: second definition in the local crate (`vmbase`)

  error: aborting due to previous error

So make it clear to Soong that libvmbase does NOT link against libstd,
so that it reports the actual vmbase build errors instead.

Test: TH
Change-Id: I812e67ed2d300e39e12084306bbf1d7047ed3ae6