Stop using TextCodebase to store kotlinStyleNulls
The `TextCodebase.kotlinStyleNulls` property is only used by `ApiFile`
during parsing. It does so because `ApiFile` is an object not a class
and so has nowhere to store its own state.

This change switches `ApiFile` to be a class. All the publicly
accessible methods are moved into a companion object so the API does
not change.

Bug: 286023667
Test: ./gradlew
Change-Id: Iadb7742a8216c95e394754ca5021393af3274002