liblog output in host now look similar to the logcat output on device
Previously, even if you used the same liblog API to emit logs, the
output was differently formatted depending on whether the program runs
on host or on device. For example,

Device side log:
06-19 17:03:50.478  3924  3924 I mytag : this is the log message

Host side log:
mytag I 06-19 04:35:20    54    54] this is the log message

Not to mention that these look quite different, the host side log had
several issues: (1) timestamp fields were not aligned if tags are
different. (2) resolution of the timestamp was low (seconds), whereas it
was ms on the device.

This CL attempts to make the host side log to look like the device side
log, by re-ordering the fields and increase the timer resolution.

Bug: 287910948
Test: see the host logs

Change-Id: I8189501d123fb79a255e17f675f8f6be617022e8