Add ClassResolver to resolve unknown classes in TextCodebase
Previously, the `TextCodebase` would add `WrappedClassItem`s around
empty stub `TextClassItem`s and then later on
`mergeClasspathIntoTextCodebase()` would redirect those
`WrappedClassItem`s to `PsiClassItem`s that it found on the class path.

That tightly coupled the `TextCodebase` and the `PsiBasedCodebase`
making it impossible to move them into separate libraries. This change
removes that coupling by adding a `ClassResolver` abstraction to which
the `TextCodebase` can delegate the responsibility for finding
missing classes to without depending on any specific implementation.

By default the `classResolver` is null which means `TextCodebase` just
creates empty stub `TextClassItem`. The code that currently calls
`mergeClasspathIntoTextCodebase()` after the `TextCodebase` is created
will now pass in an instance of `PsiBasedClassResolver` when they need
to resolve classes on the classpath.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: Ia9420698482e2fbb8cb33b80d409f35305b1d18e