Replace an old patch with two newer patches
This CL replaces the existing patch that adds support for
riscv64-linux-android with one that corrects the existing definitions
from upstream Rust and another that enables the new target.

Patch 44 is a pathological case for Toybox and can't be applied without
using the `-l` lazy match option.  An issue has been filed.

* https://github.com/landley/toybox/issues/440

Test: ./build.py --llvm-linkage shared --lto thin --gc-sections
Change-Id: I4e6597c757a03ee3a274fbcd818a0c61693cb1ff