Add more workarounds for issues in the rust-openssl port
This imports
https://github.com/sfackler/rust-openssl/commit/5283d7c994541a99bab9b33f809bd662a5aa47a7
from upstream. This patch is necessary to avoid holding back BoringSSL
updates in Android.

This and other patches in this repository are all symptoms of a deeper
problem: Android's rust-openssl port ignored BoringSSL's standard
practice and ignored our OPENSSL_VERSION_NUMBER define. As a result,
rust-openssl targets legacy APIs in BoringSSL by default and we must
regularly work around this.

The patch is already in an upstream rust-openssl release, but this crate
does not appear to have been updated in Android for six months. Import
the patch as a workaround for now.

Test: treehugger
Change-Id: I5e676adb97a32676da9df6140e6354004aa361a2