Group code needed to add items to TextCodebase for stub generation
Generating stubs from `TextCodebase` requires some additional items
that have been purposely excluded from the signature files to be added.
Previously, the code to do that was in two separate places that were
not suitable and so this change moves them into a better location.
Longer term this code will likely move into a sub-command specifically
for generating stubs from signature files. But that is not yet possible
due to other changes that are needed.

The reason the previous locations were not suitable:
* Calling `ApiAnalyzer.addConstructors()` from
  `SignatureFileLoader.loadFiles()` was not suitable because it should
  only have been done when generating stubs but there was no check to
  ensure that and `loadFiles()` is called (indirectly via `loadFile()`)
  from lots of places.
* Calling `TextCodebase.resolveAbstractMethods()` from `postProcess()`
  was not suitable because while it did have a check to ensure that it
  was only done when generating stubs that check added a dependency on
  `options`.

Bug: 288550547
Test: ./gradlew
Change-Id: I6ea810353af273333ad2d7bace1dadb6af0329b1