Simplify resolving of WrappedClassItem
Previously, `WrappedClassItem`s were resolved to a `PsiClassItem` by
first finding the `PsiFile` for the `PsiClass` for each
`WrappedClassItem`, and then using that list of `units` to initialize a
`PsiBasedCodebase`. The top level `PsiClassItem`s from the codebase
were then iterated over and if there was a corresponding
`WrappedClassItem` then it was modified to refer to the `PsiClassItem`.

It appears as though that was done to get the list of `PsiClassItem`
corresponding to each `WrappedClassItem` into the
`topLevelClassesFromSource` property. However, that was not strictly
necessary as in this case the `PsiClassItem`s are not treated as if
they came from the source. They are not emitted or contribute to an API
in any way, except through inheritance as they do when generating an
API from sources.

This change passes an empty `units` to the `PsiBaseCodebase` and uses
`PsiBasedCodebase.findOrCreateClass(PsiClass)` to obtain a
`PsiClassItem` for each `WrappedClassItem`.

Bug: 288550547
Test: ./gradlew
Change-Id: I7fccc47c753f6fd2ab5a17522a106ac68ac92fb2