tools: Add tools/shim
//tools is a bit of a mess of various naming schemes:
foo_bar
foo-bar
foo_bar.py
foo-bar.py
foobar

Even just considering the Python scripts. Everything that doesn't end
with .py doesn't get formatted aromatically (which is one problem)
but worse they can't easily share code - which leads to a lot of
duplication.

It would be better if they could live in //python/tools and get those
advantages but some people are used to the naming of the existing
scripts.

The idea is for a tool named /tools/foo-bar you:
- mv it to //python/tools/foo_bar.py
- symlink //tools/foo-bar to //tools/shim

//tools/foo-bar then continues to work as an alias for
//python/tools/foo_bar.py but you get the advantage of formatting
& code sharing.

Change-Id: If7ec4fddb0a613f02ddec1ad59ecddf880e4b428