Improve OOME handling and detection
Do not trigger an OOME just because we did not see much free memory
after running a GC; someone else may have used it up in the interim.

Attempt to handle OOME after the GC has shut down in a somewhat
better way. This is still not a great solution, but since real
Android apps don't shut down "normally", this is a significant
improvement for a test-only problem.

Simplify the homogeneous space compaction code. We now also
insist that after homogeneous space compaction, a certain
fraction of the heap be empty. (Modern GC configurations don't
use this code anyway.)

Add runFinalization() call to 2042-reference-processing. Without
that, and with changes to the OOME code that should have been
irrelevant, I saw it running out of memory, presumably because
reference processing was falling behind. Once the main thread is
close to running out of memory, it may not get around to
processing references anymore, making things worse.

Bug: 288982675
Bug: 288446884
Test: TreeHugger
Change-Id: I05fbdc13194f64b7ce959db0286d134a980f9fa9