Version the test info cache
Store the test info cache with a version discriminator. This enables
automatically invalidating user caches when Atest changes in an
incompatible way. For example, aog/2619290, requires users clean the
cache to pick up a performance improvement.

Note that we move the cache directory to the build output directory. We do this because this directory is periodically cleaned and don't have to worry about the files growing without bound. The files are also much smaller than typical build output and less of an issue.

Bug: 285399002
Test: atest atest_unittests --host --no-bazel-mode ;\
      m atest ; atest-dev atest_unittests --host --no-bazel-mode ; \
      atest-dev atest_unittests --host --no-bazel-mode
Test: rm-rf $ANDROID_PRODUCT_OUT ; atest-dev atest_unittests --host \
      --no-bazel-mode -c
Test: rm-rf $ANDROID_PRODUCT_OUT ; atest-dev atest_unittests --host \
      --no-bazel-mode

Change-Id: I373e999b1e5e577d07b5c613073cce918b06546c