Extract ConvertFile and the process() method
The `ConvertFile` data class was defined in `Options` and the code for
processing it was defined in a for loop in `Driver.kt`. This change
extracts the class into its own file and extracts the code for
processing it into the `process()` extension function of `ConvertFile`.

In preparation for the follow up change to use it for a conversion
sub-command.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: Ib4b6a8e15afa4c30860b7853b301610bfcb671ae