Avoid using options in SignatureFileLoader load* methods
The `Options.apiClassResolution` option is defined by Clikt so cannot
be read without first being initialized by a command. However, it does
not need initializing for commands that have a fixed resolution. This
change adds an `apiClassResolution` parameter to
`SignatureFileLoader`'s `load` and `loadFiles` method.

The new parameter defaults to `API` because using `API_CLASSPATH` only
has an effect when the returned `TextCodebase` is passed to
`mergeClasspathIntoTextCodebase()`. Irrespective of which value is
passed to the `load*()` methods the resulting `TextCodebase`s are
broadly equivalent:

* When `API` is used then the `TextCodebase` will create empty stub
  `TextClassItem`s for any classes which are not in the signature file.

* When `API_CLASSPATH` is used then the `TextCodebase` will create
  `WrappedClassItem`s that wrap an empty stub `TextClassItem` for any
  classes which are not in the signature file.

The `TextCodebase`s contain the exact same set of classes, and apart
from the type each `ClassItem` has the same behavior.

It is only when the `TextCodebase` is passed to
`mergeClasspathIntoTextCodebase()` that the `WrappedClassItem`s will be
redirected to `PsiClassItem`s and change the contents compared to the
empty stub classes.

Only those calls whose resulting `TextCodebase`s are passed to
`mergeClasspathIntoTextCodebase()` will pass in
`options.apiClassResolution` all the others will rely on the default.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: I203bbbd49077ef908b4ea96bf15cb93a0736e945