Add test corpus for merge sequence generator
Each cycle_nodes_*.bin is a serialized PartitionUpdate protobuf message.
The merge_sequence() field of PartitionUpdate() stores list of merge
operations before generating merge sequence.

The merge sequence generator is heuristic based, so there should not be
a fixed expectation in test cases. Removed hard coded sequence
expectations in test, and fallback to verifying that a sequence is
correct.

Test: th
Bug: 289236484
Change-Id: I0d16f5563952535816f0747883f0dd6b210649ab