[scudo] Use getMonotonicTimeFast for tryLock.
In tryLock, the Precedence value is set using the fast time function
now. This should speed up tryLock calls slightly.

This should be okay even though the value is used as a kind of random
value in getTSDAndLockSlow. The fast time call still sets enough bits
to avoid getting the same TSD on every call.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D154039

GitOrigin-RevId: 36ca9a29025a2f678096e9545fa2ec44e8432592
Change-Id: Idba52e5f41baecd58d8bba13dea80f0d6b30253e