Add enumOption() to simplify adding Enum based options
While Clikt does have explicit support for enum base options it has a
number of limitations, it:
* Defaults to allow any case.
* Does not provide support for providing and formatting help for the
  individual values.
* Uses the default `toString()` representation when displaying the
  default value, it does not map the enum value to the command line
  option.

This change adds the `enumOption()` to address those issues:
1. It defaults to lower case of the name.
2. It provides the `enumValueHelpGetter` to get the help text for each
   individual `Enum` value, e.g. from a `help` property on the value.
3. It formats the help for them using `NEL` to add line separators and
   uses markdown like syntax to identify the enum value and style it.
4. It annotates the default value with `(default)`.

The styling is a little complicated because at the time the help text
for the enum is generated (inside the `enumOption()` method) it does
not have access to the `terminal`. So, this wraps the enum value that
needs styling in markdown like syntax and then when styling the text in
`HelpFormatter` it scans the help text for the markdown syntax and
replaces it with terminal styling characters.

The use of `NEL` also causes some lines to have trailing white space
which causes problems when uploading so this removes that too.

It uses the new `enumOption()` method to simplify and improve the
current enum based options.

Bug: 286023667
Test: ./gradlew
      ./gradlew installDist
      export PATH=/usr/lib/jvm/java-17-openjdk-amd64/bin:$PATH
      ../../out/metalava/install/metalava/bin/metalava -h
Change-Id: I63def63a186313e90a0821a2df5fac2785a5083c