Improve error handling when failures occur during option parsing
Previously, if an error occurred before the `CommonOptions` was
initialized then any attempt to print the help or usage text would
cause another error to be thrown which would hide the root cause. That
happened because it would try and access the `CommonOptions.terminal`
property which would not be initialized and so would fail.

This change protects against that by renaming `terminal` to
`unsafeTerminal` and adding a new `terminal` property that is lazily
initialized and will ignore exceptions thrown while trying to access
`unsafeTerminal`. Default handling was moved into the `terminal`
property so that it was consistent whether the `unsafeTerminal` was
initialized or not. Removing the `default` from `unsafeTerminal`
caused it to be of type `Terminal?` with `null` being not set. The
lazy initializer for `terminal` handles an exception and null in the
same way.

It also adds a test which will ensure that when an error does occur
before `CommonOptions` is properly initialized that the
`unsafeTerminal` was not set. That is necessary because the window
within which this problem occurs is very small and it is quite possible
that a refactoring could narrow that window to the point that the test
was no longer testing the error handling in `terminal` which could
lead to that error handling degrading.

Bug: 289264921
Test: ./gradlew
Change-Id: Ifc519a6fe46bfb7400d934c5630b5fdea8f86704