Add Item.location()
Previously, the `psi()` method was used to retrieve the `Location`
for `PsiItem`s. This change adds an `Item.location()` method to
provide access to the `Location` directly and provides an
implementation for `PsiItem` and `TextItem`.

The `TextItem` implementation only provides a `baselineKey`. A follow
up change will use the `TextItem.position` to set `path` and `line` as
that has a big, albeit superficial, impact on tests.

This also moves `getBaselineKey(Item)` to `Location` for use by the
implementations of `Item.location()`.

Bug: 288398409
Test: ./gradlew
Change-Id: Ic894c90e4963520b535268c97e678435dacc1600