Add full support for TextItem.location()
Previously, `TextItem.location()` only exposed the `BaselineKey`. This
change updates that to use the information from `TextItem.position` to
set the `path` and `line` and modifies `Reporter` to use it when
reporting the source location.

That causes a number of tests to fail because previously, when
reporting issues with a `TextItem` the path would not be made relative
to `rootFolder` (the test temporary directory root). As a result tests
for such issues relied on the `DriverTest.cleanupString()` method to
remove the temporary directory from the test output. The consequence of
that was those tests would end up with `TESTROOT` in the output.

Bug: 288398409
Test: ./gradlew
Change-Id: Ib6760049ddf9eed2b8a2d1b3ded74e72a9548249