Autoupdate of 2 submodules
*packages/modules/Virtualization updated to 46d77fc12904570ecda153843eb38d81545e57d7
Merge changes from topic "microdroid_log"

* changes:
  Add TestApis for using console input
  Support console input to VM

*system/sepolicy updated to bd1be6c554aa065b9df8f40ed905d339cc32e754
Allow microdroid_payload to read /dev/console

The first serial device of the VM can be made bi-directional. When it is
used as an output device, it's via /dev/kmsg. microdroid_payload already
has a write access to it. When it is used as an input device, it's via
/dev/console. Grant microdroid_payload read access to the device.

Bug: 263360203
Test: atest MicrodroidTestApp:com.android.microdroid.test.MicrodroidTests#testConsoleInputSupported
Change-Id: Ief039d06ffbddee1e254d662a6c1f321a607d5f5