Add test to show broken behavior of baseline in autofill
A previous change broke the AndroidX autofill build as it used an
invalid baseline key. The problem occurs on a setter method that
uses a generic type variable (e.g. as its return type) which is
inherited by a class that binds that type variable to a specific
type. That causes a synthesized method to be created in the
sub-class that does not have an associated file.

The problem was that if a location could not be found then it
would not provide a `BaselineKey` either and so would default
to "?" which is what is seen in the test that is added in this
change. A follow up change will fix the test.

Bug: 288398409
Test: ./gradlew
Change-Id: I3b37cd3c3c7a960b57d576ffb0e17421002cc3f2