ui: Move rafScheduler out of globals
RafScheduler is one of the few things that does deserve to be its own
static singleton as it's a) legitimately used from almost everywhere,
b) it has a lifetime matching that of the application.

We still want to get rid of globals though so move it out from globals
and update all the imports.

This reduces the:
- number of files that import globals, 142 -> 120
- reduces the number of uses of globals, 1109 -> 956

Remaining top offenders:
  ...snip...
  19 globals.logging
  20 globals.getCurrentEngine
  21 globals.publishRedraw
  24 globals.makeSelection
 103 globals.frontendLocalState
 196 globals.dispatch
 344 globals.state

Change-Id: I9333a6e9ee943bb820003e7391bb732eb3a9b1bd