Ensure that cache files are valid
Atest sometimes exits without properly writing the cache file. This can
happen because we don't wait for the saving process or because the
writing operation is aborted midway.

This change addresses the situation by ensuring that Atest waits for the
cache file to be completely rewritten before exiting. We also write the
cache file to a temporary location and move it to the target path. This
ensures that the operation is atomic.

Note that the DB version number is incremented to force regeneration for
users that may have an invalid partial cache.

Bug: 289882784
Test: atest-dev atest_unittests
Test: atest-dev HelloWorldHostTest -m
Test: atest-dev HelloWorldHostTest
Test: atest-dev HelloWorldHostTest -m --no-sqlite-module-cache
Test: atest-dev HelloWorldHostTest --no-sqlite-module-cache
Change-Id: I60b753e4e2a2fe7bf79364711d9f69762c004108