Add Item.implicitNullness() and override in MemberItem
The `AnnotationItem.getImplicitNullness(Item)` method breaks the model
abstraction by down casting to various model specific types. This
change adds an `implicitNullness()` method to `Item` and starts the
process of moving the item specific behavior down into those methods.

It adds an override for the `implicitNullness()` method to `MemberItem`
and move the logic that marks annotation members as non-null.

Bug: 289919515
Test: ./gradlew
Change-Id: I5ef213e8a7d769ac13b24dc56a57378c346b5e24