Restructure AnnotationItem.getImplicitNullness()
The method actually consists of a number of mutually exclusive checks
but it is not clear from the way that the code is structured. That
would make it difficult to extract the logic and implement it on the
appropriate `Item` sub-class, including, where necessary, on a
`PsiItem` sub-class.

This change restructures the method to make the mutually exclusive
nature of the checks clearer. It also moves the check for companion
objects into the `FieldItem` check as that is how companion objects
are represented in the models.

Bug: 289919515
Test: ./gradlew
      # On androidx-main/androidx_with_metalava target
      # Merged into aosp/master and run `m checkapi`
Change-Id: I0cb85d46199a3e6f18f670e774139491effa9489