Override implicitNullness() in FieldItem and PsiFieldItem
Moves the logic related to `FieldItem` and `PsiFieldItem` from
`AnnotationItem.getImplicitNullness()` to overrides of the
`implicitNullness()` method on the appropriate `Item` sub-class.

The order of the field checks are preserved, which is why the
the companion object check is performed before calling the super
class's implementation and the other checks are performed
afterwards.

Bug: 289919515
Test: ./gradlew
Change-Id: I8f0c73ee40de70f453d72a759d1a32aa6a915db8