hals: Add lights service implementation for Cuttlefish
This CL adds a custom implementation of the Lights HAL service for
Cuttlefish and changes the buildsystem to point at it. We are starting
this as a copy of the example service provided by Android (written in
Rust) with only the service and package names modified.

There are no behavior changes at this time, but future work will add a
VSOCKs backend to this service and implement a simple UI to reflect the
lights state on the Cuttlefish frontend.

Bug: 286106270
Change-Id: Iece786e2720b320a9c5cd75a4975a09680a6e8a3