Extract DefaultAnnotationManager from AnnotationItem
The `AnnotationItem.mapName` and `computeTargets` methods depend on the
`Options` class and `options` global which would prevent the model from
being extracted into its own library.

This change extracts the functionality out into the `AnnotationManager`
interface and the `DefaultAnnotationManager` implementation. That is in
preparation for follow up changes which will refactor the callers of
the `mapName` and `computeTargets` to use an `AnnotationManager`
instance they are supplied rather than calling the global methods.

The default instance of `DefaultAnnotationManager` is obtained from
`Options.annotationManager`.

It also adds an `annotationManager` property to `Codebase` and
`DefaultCodebase` that for now defaults to `defaultAnnotationManager`
and cleans up usages of `mapName()` in the `...model` package.

Bug: 289919515
Test: ./gradlew
Change-Id: I25ffccac15a6712035bbb95fb8d9fc8e5e8ddf4b