Improve stability of screenshot tests
This introduces a screenshot and getscreenshot fixture that will save
the obtained screenshot to the logdirectory, which can be used to visual
debug issues if we encounter them.

We also improve the power test to use pulling screenshots instead of
streaming them, to make sure we do not have to rely on the emulator
screenshot notification mechanism.

Change-Id: I8e54c18da2f1ceef81829298a9874295aa0787b1