Only look at resolved classes for generating a runtime image.
There could be a race where we first see a loaded class (and its
interface/super isn't in the class table yet), but we then see the class
as resolved and emit it. In such a situation, we would emit the super
class and interfaces, but forget to add them to the class table.

For simplicity, we now only collect classes which are resolved.

Test: test.py (unfortunately, difficult to unit test due to the race)
Bug: 287502985
Change-Id: Icfc578b1e28d2320046cb4e12efc1810ee9074c7