Remove DefaultAnnotationManager dependency on ApiPredicate
Adds `apiPredicate: Predicate<Item>` to the config and initialize that
as `ApiPredicate()` in `options`. The default value, which is only used
for unit tests and a GTS test (via
`ApiFile.parseFile(filename, apiText)`), always returns `true`. That is
ok as both types of test use it to load an API signature file which has
already been filtered.

Bug: 289919515
Test: ./gradlew
Change-Id: I8f0de270bdf3dea0f577f0e4a8c5e1f2cce4fbe5