Keep Exceptions attributes for RkpdApp
This is required when using Mockito's thenThrow with checked exceptions.
When R8 full mode is enabled, an explicit keep rule is required to keep
such attributes. For now, we use the minimal keep rule for mocked class
methods, but this could be relaxed for packages if that would simplify
testing going forward.

Bug: 215530220
Bug: 289322782
Change-Id: I2f558bca60bddfa666b2bf3969c83a6e37ef3525
Test: atest RkpdAppGoogleUnitTests