Force DNS lookup when a new IP version is available
This commit updates IKE so that it will forcibly do DNS
lookup when the supported IP versions of the cached remote
addresses are different from the local addresses during
migration. Note a NAT64 IPv6 address will not be considered
as a valid cached address because it might be out-of-date.
Previously, DNS will be skipped in cases above. It might cause
IKE to be stuck on the less preferred IP version when the
preferred version comes up late.

Bug: 290396469
Test: atest FrameworksIkeTests; atest CtsIkeTestcases
Change-Id: Ifb3f8dd4a903cb52fb984fe60f7e38be5e0fef5f