Remove DocLevel related command line options
The command line options that determine the value of the
`options.docLevel` property do not seem to be used. This change removes
support for them and removes the property.

The `checkLevel()` method is the only user of the `options.docLevel`
property. This change rewrites it to behave as it would when the
`docLevel` is set to its default value of `PROTECTED`, i.e. the method
returns `true` if the item is not hidden or removed and is either
`public` or `protected`.

Bug: 287241408
Test: ./gradlew
Change-Id: I6c85bd88ec7fb5dccc9fda673c460c6f3063cdb1