Don't use timestamp counters on ARM32
Not all arm32 have access to timestamp counters. Even on ARMV8
processors where timestamp counters are available, they are not
accessible from user space always when running in 32-bit mode. There is
no easy way to check if the timestamp counters are supported. So for now
just fallback to clock_gettime.

Bug: 289178149
Test: art/test.py --trace --32
Change-Id: I8406f3a99a178e6806e2dc762662a7f5f0e538ea