aidl: load&validate no mutate previously loaded
The AIDL compiler does relatively few passes on its input. It is not
a fully multipass design. The original design was singlepass, and
all modifications were made to the tree at once.

As multiple passes have been added, IterateTypes was introduced in
AidlTypenames that allows you to iterate over all types. However,
this also iterates over previously loaded types.

However, as chance has it, IterateTypes was used to add meta
methods to AIDL interfaces. Because of this, if load_and_validate_aidl
was called re-using AidlTypenames, then it would add meta methods
again to already existing methods.

dump_api worked anyway because it did not specify the version or
hash file options when dumping APIs.

However, a visitor is a better pattern for this, because we can instead
explicitly only add methods to those types which are currently being
loaded, without modifying earlier types.

This CL keeps a const_cast that we originally had, but we could
get rid of it by introducing mutable variations of all the visited
functions.

Fixes: 290685834
Test: build, aidl_unittests
Change-Id: Ib79eff71b3b8b775af7457e2e66c01329d9b9c0d