Don't include configs with same name, diff content
We may have configs with same name, but with different content. for
example, if they are host and device variants of the same test config.
prioritize-host-config flag is used to decide which config to select if
multiple configs exist, by reordering the test folders mainly. returning
duplicates breaks this logic. So, adding this logic inside a flag only.
this is still needed for validating all configs, including the
duplicates.

Bug: 286559640
Test: Presubmit
Change-Id: I97d06e2558fc0783dc17c1aea4980e72f2ee9434