Move Item.markRecent() to extension function in NullnessMigration.kt
The `Item.markRecent()` method uses the `findNullnessAnnotation()`
function from `NullnessMigration.kt` which should not be part of the
model. So, this change moves the `Item.markRecent()` method out of
the model package.

An alternative solution would have been to move
`findNullnessAnnotation(Item)` to `Item.findNullnessAnnotation()` but
this is better because the `Item.markRecent()` method is not a
generally useful method, is only used by a couple of classes outside
the model and the simpler we keep the model the easier it will be to
maintain.

Given Kotlin's support for extension functions there is no need to
pollute the model classes with task specific classes and members.

Bug: 287241408
Test: ./gradlew
Change-Id: I83041b3ca5e9f86be0f49e6fdd010bd522e037d2