Move TypeItem.internalName() to JvmHelpers.kt
The `TypeItem.internalName()` calls `TypeItem.toSlashFormat()` which in
turn depends on a lint class `ClassContext`. The model package cannot
depend on lint classes so this moves the method out.

It is possible, maybe even likely, that the work to improve type
handling in Metalava (b/111253910), will be able to implement the
`internalName()` method better than the current implementation (which
relies on heuristics implemented by `ClassContext`) and so it may make
sense to move the method back.

Bug: 287241408
Test: ./gradlew
Change-Id: I3db9a8f7ae06c025eac4a9e23a4ba7a026e75b7a