Set global limiter from INT_MAX to 2500 by default
Although the global limiter is currently pushed by server, many OEM
devices do not login to the Google account to get the update before
auto testing. This commit set the default value to 2500 in module code.
This value is safe as google server has pushed the same value four
months ago and no issue was reported.

From statistical data, the issue this commit tries to fix happens only
in OEM labs with extreme conditions. The issue is not observed in the
field. So this is a fix for OEM lab testing only.

Bug: 291016548
Test: atest
Change-Id: If8d7ac9474e6682eb18659547ae7beca8f03a8f6