Improve consistency of file based options/arguments
Previously, the `Options` class used its own methods for converting a
string option/argument into a file and the Clikt based sub-commands
used Clikt's own conversion. That caused problems when migrating to use
the sub-commands as there were some slight differences in behavior.

The main difference was in the handling of an option representing a new
file. `Options` would make sure that the parent directory existed,
creating it if it did not, whereas Clikt would not. Some uses of the
command expected that behavior. There are a number of other differences
such as handling of `~/` and `@` prefixed files that could cause
problems in future.

This change creates Clikt based support for new and existing files that
reuses the `Options` functions for `String -> File` conversions to make
it easier to migrate uses of metalava to Clikt base sub-commands.

Bug: 286023667
Test: ./gradlew
Change-Id: I9087470b52770da7fe2b985c7ee889807ed822ad