Replace the tristate UFFD GC P/H flag with a boolean one.
Our server config defines the UFFD GC P/H flag as a boolean one, while
it's actually implemented as a tristate one. It's very unfortunate that
we can't change the server config. Therefore, this CL retires the
tristate flag, and creates a real boolean flag instead.

After this change, we'll lose the ability to override a build-time
"true" with a "false" from the server, but that's okay.

Bug: 291060471
Test: atest odsign_e2e_tests_full
Test: (on a V device) -
  1. Boot the device and see "CollectorTypeCMC GC".
  2. adb shell device_config put runtime_native_boot enable_uffd_gc_2 false
  3. Reboot.
  4. See no re-compilation.
  5. Still see "CollectorTypeCMC GC".
Test: (on a U device) -
  1. Boot the device and see "CollectorTypeCC GC".
  2. adb shell device_config put runtime_native_boot enable_uffd_gc_2 false
  3. Reboot.
  4. See no re-compilation.
  5. Still see "CollectorTypeCC GC".
Change-Id: I716ffa3406feda4d7790a7ca8c331169f09d0baa