Simplify handling of DeprecatedForSdk
Previously, the `DeprecatedForSdk` annotation was mapped to the
standard `Deprecated` annotation and then any code that needed to treat
`DeprecatedForSdk` as different to `Deprecated` had to check the
`originalName`. It also required a special `findExactAnnotation()`
method.

This change passes the `DeprecatedForSdk` annotation through intact and
treats it as any normal annotation. It also gives it an empty set of
output targets (just like `kotlin.Deprecated`) as if it does affect the
output file it does so by being converted to a standard `Deprecated`
annotation.

FTR the conversion to `Deprecated` is done via the `Item.deprecated`
property. The `PsiModifierItem` will filter out a `DeprecatedForSdk`
annotation if it is not suitable for the output API surface. The
`ApiAnalyzer` checks to see if an `Item` has a `DeprecatedForSdk`
annotation and if it does then it sets `Item.deprecated = true` and
adds an `@deprecated` comment.

Bug: 291562733
Test: ./gradlew
Change-Id: I25aa94a58a1ce154613f4245dfcbf8007629fcfc