Use static_cast to disambiguate assembler instructions.
Because of misfeature of Itanium C++ ABI we couldn't just use
MacroAssembler to static cast these references if we want to use
them as template argument:
  https://ibob.bg/blog/2018/08/18/a-bug-in-the-cpp-standard/

Thankfully there are no need to use the same name for MacroInstructions
since we may always rename these.

But for assembler we need to use actual type from where these
instructions come from!

Bug: 278812060

Test: m berberis_all

Change-Id: Ie2a4a8fefb10836b8793c80b8d9d13dba655dfa7